Factor local variable table functions and clean up the convert_method class.
