+++ src/java_bytecode/Makefile
@@ -6,7 +6,7 @@ SRC = java_bytecode_language.cpp java_bytecode_parse_tree.cpp \
       java_root_class.cpp java_bytecode_parser.cpp bytecode_info.cpp \
       java_class_loader.cpp jar_file.cpp java_object_factory.cpp \
       java_opaque_method_stubs.cpp java_bytecode_convert_method.cpp \
-      java_pointer_casts.cpp
+      java_pointer_casts.cpp java_local_variable_table.cpp
 
 INCLUDES= -I ..
